containers: docker-compose.fbdev-simple-test.yml: added an example docker compose demonstrating two apps built in a multi-build flow, statically and dynamically. The repository they pull from is private and for now it will remain so (I simply don't want to maintain those example apps), and if someone knows if there is a way to have a docker push to ghcr.io  without going through github workflows (or at least do the first creation, I think subsequent pushes should work), please do tell me, and I may change this. It's simply a matter of time
The containers themselves:
- dynamic - heavy - has all of glibc dependencies needed for rust - taken
  from rust:slim which is not so slim
- static - tiny

Both are included in the docker-compose file simply for demonstration,
but for this particular example, the static variation is all you need,
and is 2 orders of magnitudes smaller than the dynamic one. The apps
there are identical, only linked differently. Two apps are presented in
each container, conveniently depicting the programming language in which
they were made:
- /artifacts/c/fb - draws to FB. Input to change frame  (if not run with
 timed argument) is from stdin
- /artifacts/rust/fb - draws to FB ad switches every second if --timed
  argument is provided. If --input argument is provided - changes frames
upon an input event (e.g. /dev/input/eventX) - which is useful for
testing touch screens, keyboards, etc.